requirements.txt
# requirements.txt - ATUALIZADO COM FERRAMENTAS DE TESTE
fastapi==0.111.0
uvicorn[standard]==0.29.0
beanie==1.26.0
motor==3.4.0
pymongo==4.6.2
python-dotenv==1.0.1
google-generativeai==0.5.4
requests==2.32.3
pydantic==2.7.1

# Ferramentas de Teste
pytest
httpx
pytest-asyncio

# Ferramenta de Frontend
streamlit
pandas